requirements.txt
django==4.2
djangorestframework
psycopg2
pillow
python-decouple
smsir-python
djangorestframework-simplejwt
requests
drf-yasg    
zarinpal-payment-integration
#zarinpal 
django-cors-headers
instagrapi==2.1.2
pydantic==2.7.1